edge-client/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
psutil==5.8.0
python-dotenv==0.15.0
python-engineio==3.14.2
python-socketio==4.6.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
websocket-client==0.58.0
